--- Machete.Api/Program.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-
-namespace Machete.Api
-{
-    public class Program
-    {
-        public static void Main(string[] args) =>
-            CreateWebHostBuilder(args).Build().Run();        
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
-    }
-}
