--- Machete.Api/appsettings.json
@@ -1,16 +1,0 @@
-{
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost,1433; Database=machete_db; User=SA; Password=passw0rD!;",
-    "ReadOnlyConnection": "Server=localhost,1433; Database=machete_db; User=readonlylogin; Password=@testPassword1;"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "JwtIssuerOptions": {
-    "Issuer": "machete",
-    "Audience": "machete-ui-local-embedded"
-  },
-  "AllowedHosts": "*"
-}
