--- Machete.Web/StaticConfiguration.cs
@@ -1,135 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using Machete.Data;
-using Machete.Data.Infrastructure;
-using Machete.Data.Repositories;
-using Machete.Service;
-using Machete.Web.Helpers;
-using Machete.Web.Maps;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Hosting.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Localization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Razor;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileProviders;
-
-namespace Machete.Web
-{
-    /// <summary>
-    /// A class containing builder pattern extension methods for the IServiceCollection and IApplicationBuilder
-    /// classes, which are 
-    /// </summary>
-    public static class StaticConfiguration
-    {
-        public static PathString RootPath = PathString.Empty;
-
-        public static void ConfigureMvcBuilder(this IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseDatabaseErrorPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                app.UseHsts();
-            }
-
-            // https://docs.microsoft.com/en-us/aspnet/core/fundamentals/localization?view=aspnetcore-2.2 (Ibid.)
-            var supportedCultures = new[]
-            {
-                // Ibid. #globalization-and-localization-terms
-                // https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
-                // https://en.wikipedia.org/wiki/ISO_3166-1
-                new CultureInfo("en-US"),
-                new CultureInfo("es-US"),
-                // we use es-US because we are not fully equipped to support international dates.
-            };
-            app.UseRequestLocalization(new RequestLocalizationOptions
-            {
-                DefaultRequestCulture = new RequestCulture("en-US"),
-                // Formatting numbers, dates, etc.
-                SupportedCultures = supportedCultures,
-                // UI strings that we have localized.
-                SupportedUICultures = supportedCultures
-            });
-            // the preceding will attempt to guess the user's culture. For several reasons that's not what we want.
-            // Ibid. #set-the-culture-programmatically
-
-            app.UseCors("AllowAllOrigins"); // TODO
-
-            app.UseHttpsRedirection();
-
-            //app.UseStaticFiles(); // ?
-            app.UseStaticFiles("/Content");
-            app.UseStaticFiles(new StaticFileOptions
-            {
-                FileProvider = new PhysicalFileProvider(
-                    Path.Combine(Directory.GetCurrentDirectory(), "Content")),
-                RequestPath = "/Content"
-            });
-
-            app.UseCookiePolicy();
-
-            app.UseAuthentication();
-
-            app.UseMvc(routes =>
-            {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Account}/{action=Login}/{id?}");
-                routes.MapRoute(
-                    name: "V2",
-                    template: "V2/{*url}",
-                    defaults: new {controller = "V2", action = "Index"}
-                );
-            });
-
-            app.UseDirectoryBrowser(new DirectoryBrowserOptions
-            {
-                FileProvider = new PhysicalFileProvider(
-                    Path.Combine(Directory.GetCurrentDirectory(), "Content")),
-                RequestPath = "/Content"
-            });
-
-            app.UseMvcWithDefaultRoute();
-        }
-
-        public static IWebHost CreateOrMigrateDatabase(this IWebHost webhost)
-        {
-            using (var scope = webhost.Services.CreateScope())
-            {
-                var context = scope.ServiceProvider.GetService<MacheteContext>();
-                context.Database.Migrate();
-                MacheteConfiguration.Seed(context, webhost.Services);
-            }
-
-            return webhost;
-        }
-
-        private static string GetControllerNames()
-        {
-            var controllerNames = Assembly.GetCallingAssembly()
-                .GetTypes()
-                .Where(x =>
-                    x.IsSubclassOf(typeof(ControllerBase)) &&
-                    x.FullName.StartsWith(MethodBase.GetCurrentMethod().DeclaringType.Namespace + ".Controllers"))
-                .ToList()
-                .Select(x => x.Name.Replace("Controller", ""));
-
-            return string.Join("|", controllerNames);
-        }
-    }
-}
