+++ Machete.Web/appsettings.json
@@ -8,5 +8,9 @@
       "Default": "Warning"
     }
   },
+  "JwtIssuerOptions": {
+    "Issuer": "machete",
+    "Audience": "machete-ui-local-embedded"
+  },
   "AllowedHosts": "*"
 }
