+++ Machete.sln
@@ -11,8 +11,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Machete.Service", "Machete.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Machete.Test", "Machete.Test\Machete.Test.csproj", "{69EC5263-745F-40CB-9E53-FC091FE1619C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Machete.Api", "Machete.Api\Machete.Api.csproj", "{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -336,63 +334,5 @@ Global
 		{69EC5263-745F-40CB-9E53-FC091FE1619C}.UAT-test|Any CPU.Build.0 = Debug|Any CPU
 		{69EC5263-745F-40CB-9E53-FC091FE1619C}.Debug-no_nodejs_build|Any CPU.ActiveCfg = Debug|Any CPU
 		{69EC5263-745F-40CB-9E53-FC091FE1619C}.Debug-no_nodejs_build|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.StageProd|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.StageProd|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.StageTest|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.StageTest|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.CasaLatina|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.CasaLatina|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Cloud|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Cloud|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.MacheteTest|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.MacheteTest|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.WorkersJustice|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.WorkersJustice|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.PortlandVoz|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.PortlandVoz|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Southside|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Southside|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.DonBosco|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.DonBosco|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.CentroHumanitario|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.CentroHumanitario|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.SantaCruz|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.SantaCruz|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.SanFrancisco|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.SanFrancisco|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Pomona|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Pomona|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Pasadena|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Pasadena|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Nice|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Nice|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.MountainView|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.MountainView|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Hayward|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Hayward|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Graton|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Graton|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.ElCentro|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.ElCentro|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Concord|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Concord|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Congreso|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Congreso|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Alabama|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Alabama|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.WindofSpirit|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.WindofSpirit|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Cel|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Cel|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.localIIS|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.localIIS|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.UAT-test|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.UAT-test|Any CPU.Build.0 = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Debug-no_nodejs_build|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50EF244-42A3-4E92-9EAF-0A24AEBB6BBC}.Debug-no_nodejs_build|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 EndGlobal
